+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - **Breaking Change**: replace PCRE2 with own regex engine
     - now regex syntax and semantics are compatible with ECMAScript regular expression (ES2025)
     - enable Unicode mode by default
+- optimize simple case folding
 
 #### Builtin